set: add missing set/unset support for NFT_SET_ATTR_DATA_[TYPE|LEN]
While at it, use fixed length uint32_t instead of size_t.

+++ src/internal.h
@@ -42,9 +42,9 @@ struct nft_set {
 	char			*table;
 	char			*name;
 	uint32_t		key_type;
-	size_t			key_len;
+	uint32_t		key_len;
 	uint32_t		data_type;
-	size_t			data_len;
+	uint32_t		data_len;
 	struct list_head	element_list;
 
 	uint32_t		flags;
